# support/design/openapi-fixture.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "neuradex",
    "description": "A small API service implementing tools for LLM agents: fetching pages (`/v1/fetch`, `/v1/peek`) and searching the web (`/v1/search`). Every error response — including router-level 404 and 405 — is rendered as `{\"error\": {\"type\", \"message\"}}`; see the `ErrorBody` schema.",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/fetch": {
      "get": {
        "tags": [
          "fetch"
        ],
        "summary": "Fetch a page.",
        "description": "Fetches the page and returns its document-head metadata, body, and request metrics. Up to `redirects` redirects are followed by hand, and each hop can be reported through `include=redirects`. The body is decoded as UTF-8 text with invalid sequences replaced; binary content types are rejected, and bodies beyond 25 MiB are cut short with `truncated: true`. The request and response headers are returned when `include=headers` is requested.",
        "operationId": "fetch",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "description": "The absolute URL of the page to fetch.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "example": "https://example.com"
          },
          {
            "name": "redirects",
            "in": "query",
            "description": "The maximum number of redirects to follow; defaults to 5.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5,
              "maximum": 5,
              "minimum": 0
            },
            "example": 2
          },
          {
            "name": "include",
            "in": "query",
            "description": "The optional data to include in the response; may be repeated and comma-separated.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": true,
            "example": [
              "redirects"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "The page was fetched",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchResponse"
                }
              }
            }
          },
          "400": {
            "description": "A request parameter is missing, malformed, or out of range.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "The JSON body of an error response.",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorDetail",
                      "description": "The error details."
                    }
                  },
                  "examples": [
                    {
                      "error": {
                        "message": "the timeout parameter must be between 1 and 30 seconds",
                        "type": "invalid_param"
                      }
                    }
                  ]
                }
              }
            }
          },
          "415": {
            "description": "The target response has a content type that cannot be returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "The JSON body of an error response.",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorDetail",
                      "description": "The error details."
                    }
                  },
                  "examples": [
                    {
                      "error": {
                        "message": "the timeout parameter must be between 1 and 30 seconds",
                        "type": "invalid_param"
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "An internal, unexpected failure.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "The JSON body of an error response.",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorDetail",
                      "description": "The error details."
                    }
                  },
                  "examples": [
                    {
                      "error": {
                        "message": "the timeout parameter must be between 1 and 30 seconds",
                        "type": "invalid_param"
                      }
                    }
                  ]
                }
              }
            }
          },
          "502": {
            "description": "The upstream request failed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "The JSON body of an error response.",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorDetail",
                      "description": "The error details."
                    }
                  },
                  "examples": [
                    {
                      "error": {
                        "message": "the timeout parameter must be between 1 and 30 seconds",
                        "type": "invalid_param"
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/peek": {
      "get": {
        "tags": [
          "peek"
        ],
        "summary": "Fetch a page's head metadata.",
        "description": "Fetches the page and returns only its document-head metadata. The response is streamed through the head parser and the download is aborted as soon as the head has been received — through `</head>` or the start of `<body>` — so the endpoint returns as early as possible. Unlike `/v1/fetch`, the body is not returned and `include=headers` is not supported.",
        "operationId": "peek",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "description": "The absolute URL of the page to fetch.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "example": "https://example.com"
          },
          {
            "name": "redirects",
            "in": "query",
            "description": "The maximum number of redirects to follow; defaults to 5.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5,
              "maximum": 5,
              "minimum": 0
            },
            "example": 2
          },
          {
            "name": "include",
            "in": "query",
            "description": "The optional data to include in the response; may be repeated and comma-separated.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": true,
            "example": [
              "redirects"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "The page head was fetched",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeekResponse"
                }
              }
            }
          },
          "400": {
            "description": "A request parameter is missing, malformed, or out of range.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "The JSON body of an error response.",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorDetail",
                      "description": "The error details."
                    }
                  },
                  "examples": [
                    {
                      "error": {
                        "message": "the timeout parameter must be between 1 and 30 seconds",
                        "type": "invalid_param"
                      }
                    }
                  ]
                }
              }
            }
          },
          "415": {
            "description": "The target response has a content type that cannot be returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "The JSON body of an error response.",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorDetail",
                      "description": "The error details."
                    }
                  },
                  "examples": [
                    {
                      "error": {
                        "message": "the timeout parameter must be between 1 and 30 seconds",
                        "type": "invalid_param"
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "An internal, unexpected failure.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "The JSON body of an error response.",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorDetail",
                      "description": "The error details."
                    }
                  },
                  "examples": [
                    {
                      "error": {
                        "message": "the timeout parameter must be between 1 and 30 seconds",
                        "type": "invalid_param"
                      }
                    }
                  ]
                }
              }
            }
          },
          "502": {
            "description": "The upstream request failed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "The JSON body of an error response.",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorDetail",
                      "description": "The error details."
                    }
                  },
                  "examples": [
                    {
                      "error": {
                        "message": "the timeout parameter must be between 1 and 30 seconds",
                        "type": "invalid_param"
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "type": "object",
        "description": "The details of an error response.",
        "required": [
          "type",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "The human-readable error message."
          },
          "type": {
            "type": "string",
            "description": "The machine-readable error type."
          }
        }
      },
      "FetchResponse": {
        "type": "object",
        "description": "The response of the fetch endpoint.",
        "required": [
          "url",
          "metadata",
          "body",
          "truncated",
          "metrics"
        ],
        "properties": {
          "body": {
            "type": "string",
            "description": "The response body, decoded as UTF-8 with invalid sequences replaced."
          },
          "metadata": {
            "$ref": "#/components/schemas/PageMetadata",
            "description": "The metadata extracted from the document head."
          },
          "metrics": {
            "$ref": "#/components/schemas/Metrics",
            "description": "The request metrics and stats."
          },
          "request_headers": {
            "type": [
              "object",
              "null"
            ],
            "description": "The request headers, when requested through `include=headers`.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "response_headers": {
            "type": [
              "object",
              "null"
            ],
            "description": "The response headers, when requested through `include=headers`.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "truncated": {
            "type": "boolean",
            "description": "Whether the body was cut short by the size limit."
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "The requested URL."
          }
        },
        "examples": [
          {
            "body": "<!doctype html>\n<html>…</html>",
            "metadata": {
              "canonical": "https://example.com/",
              "description": "A small API service implementing tools for LLM agents.",
              "og": {
                "site_name": "Maero"
              },
              "other": {
                "viewport": "width=device-width"
              },
              "title": "Maero",
              "twitter": {}
            },
            "metrics": {
              "bytes_read": 5123,
              "content_length": 5123,
              "content_type": "text/html",
              "final_url": "https://example.com/",
              "http_version": "HTTP/2",
              "redirects": [],
              "redirects_followed": 0,
              "status": 200,
              "total_ms": 120,
              "truncated": false,
              "ttfb_ms": 87
            },
            "request_headers": {
              "accept-encoding": "gzip, deflate, br, zstd",
              "user-agent": "neuradex/0.1"
            },
            "response_headers": {
              "content-length": "5123",
              "content-type": "text/html; charset=utf-8"
            },
            "truncated": false,
            "url": "https://example.com/"
          }
        ]
      },
...
      "PeekResponse": {
        "type": "object",
        "description": "The response of the peek endpoint.",
        "required": [
          "url",
          "metadata",
          "metrics"
        ],
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/PageMetadata",
            "description": "The metadata extracted from the document head."
          },
          "metrics": {
            "$ref": "#/components/schemas/Metrics",
            "description": "The request metrics and stats."
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "The requested URL."
          }
        },
        "examples": [
          {
            "metadata": {
              "canonical": "https://example.com/",
              "description": "A small API service implementing tools for LLM agents.",
              "og": {
                "site_name": "Maero"
              },
              "other": {},
              "title": "Maero",
              "twitter": {}
            },
            "metrics": {
              "bytes_read": 1024,
              "content_length": 5123,
              "content_type": "text/html",
              "final_url": "https://example.com/",
              "http_version": "HTTP/2",
              "redirects": [],
              "redirects_followed": 0,
              "status": 200,
              "total_ms": 95,
              "truncated": false,
              "ttfb_ms": 87
            },
            "url": "https://example.com/"
          }
        ]
      },
...
    }
  },
...
}